requirements.txt
streamlit
matplotlib
numpy
Pillow
fpdf